<refactor>[compute]: split VmInstanceManagerImpl 2808L into 6 SubManagers
Resolves: ZSTAC-82456
